+++ docs/main.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import VueRouter from 'vue-router'
 import App from './components/App.vue'
 import routes from './routes'
-import VueCesium from '../src/index.js'
+
 import DocPreview from './components/DocPreview.vue'
 import TextField from './components/TextField.vue'
 import VueSlider from 'vue-slider-component'
@@ -13,8 +13,9 @@ import 'material-design-icons/iconfont/material-icons.css'
 import './fonts/iconfont.css'
 import 'vue-material/dist/vue-material.css'
 import 'vue-slider-component/theme/default.css'
-Vue.use(VueMaterial)
-Vue.use(VueRouter)
+import VueCesium from '../lib/index.js'
+import '../lib/vc-navigation.css'
+import '../lib/vc-navigation-sm.css'
 Vue.use(VueCesium, {
   // cesiumPath is path of Cesium.js', for example:
   // local Cesium Build package:
@@ -31,6 +32,9 @@ Vue.use(VueCesium, {
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiI5Y2U0ZTk2Ni1jNzdkLTQ3OWYtYjVmYS0yMGM3YTk3NjgzMmUiLCJpZCI6Njk5Nywic2NvcGVzIjpbImFzciIsImdjIl0sImlhdCI6MTU0ODA1MTc0OH0.Csy6yyAnv6JSBppH0Ou3ahshqcHFEhP27iOz5gjQMEo'
 })
 
+Vue.use(VueRouter)
+
+Vue.use(VueMaterial)
 Vue.material.registerTheme({
   white: {
     primary: 'white',